refactor: move ws server to separate file
